--- python/tweedledum/BoolFunctionCompiler/types.py
@@ -1,9 +1,0 @@
-#-------------------------------------------------------------------------------
-# Part of Tweedledum Project.  This file is distributed under the MIT License.
-# See accompanying file /LICENSE for details.
-#-------------------------------------------------------------------------------
-from enum import Enum
-
-class CompilationType(Enum):
-    PHASE = 0
-    COMPUTATIONAL_BASIS = 1
